chore: Moves mesh-insights-reducer into main-overview module
